*.elf ignored
